+++ .github/workflows/x11-xcb-proto-1.13.yml
@@ -3,14 +3,14 @@ name: xcb-proto-1.13 CI
 on:
   pull_request:
     paths:
-      - 'recipes/x11/all/xcb-proto/1.13/**'
+      - 'recipes/xcb-proto/1.13/**'
       - 'build.py'
       - '.github/workflows/x11-xcb-proto-1.13.yml'
   push:
     branches:
       - master
     paths:
-      - 'recipes/x11/all/xcb-proto/1.13/**'
+      - 'recipes/xcb-proto/1.13/**'
       - 'build.py'
       - '.github/workflows/x11-xcb-proto-1.13.yml'
 
@@ -65,7 +65,7 @@ jobs:
           install: custom
           # TODO: replace when https://github.com/conan-io/conan-package-tools/issues/479 will be fixed upstream
           custom-package: git+https://github.com/trassir/conan-package-tools@fix-479-trassir
-          work-dir: sources/recipes/x11/all/xcb-proto/${{ matrix.xcb-proto_versions }}
+          work-dir: sources/recipes/xcb-proto/${{ matrix.xcb-proto_versions }}
           build-script: ../../../build.py
           compiler: ${{ matrix.compiler }}
           compiler-versions: ${{ matrix.compiler_version }}
